move analysis components to components folder
