+++ apps/web/app/(app)/environments/[environmentId]/surveys/[surveyId]/(analysis)/summary/components/SummaryPage.tsx
@@ -1,27 +1,28 @@
 "use client";
+
 import CustomFilter from "@/app/(app)/environments/[environmentId]/surveys/[surveyId]/CustomFilter";
 import SummaryHeader from "@/app/(app)/environments/[environmentId]/surveys/[surveyId]/SummaryHeader";
-import SurveyResultsTabs from "@/app/(app)/environments/[environmentId]/surveys/[surveyId]/(analysis)/SurveyResultsTabs";
-import ResponseTimeline from "@/app/(app)/environments/[environmentId]/surveys/[surveyId]/(analysis)/responses/ResponseTimeline";
+import SurveyResultsTabs from "@/app/(app)/environments/[environmentId]/surveys/[surveyId]/(analysis)/components/SurveyResultsTabs";
+import SummaryList from "@/app/(app)/environments/[environmentId]/surveys/[surveyId]/(analysis)/summary/components/SummaryList";
+import SummaryMetadata from "@/app/(app)/environments/[environmentId]/surveys/[surveyId]/(analysis)/summary/components/SummaryMetadata";
 import ContentWrapper from "@/components/shared/ContentWrapper";
 import { useResponseFilter } from "@/app/(app)/environments/[environmentId]/ResponseFilterContext";
 import { getFilterResponses } from "@/lib/surveys/surveys";
 import { TResponse } from "@formbricks/types/v1/responses";
-import { TSurvey } from "@formbricks/types/v1/surveys";
+import { TSurveyWithAnalytics } from "@formbricks/types/v1/surveys";
 import { useSearchParams } from "next/navigation";
 import { useEffect, useMemo } from "react";
 
-interface ResponsePageProps {
+interface SummaryPageProps {
   environmentId: string;
-  survey: TSurvey;
+  survey: TSurveyWithAnalytics;
   surveyId: string;
   responses: TResponse[];
   surveyBaseUrl: string;
 }
 
-const ResponsePage = ({ environmentId, survey, surveyId, responses, surveyBaseUrl }: ResponsePageProps) => {
+const SummaryPage = ({ environmentId, survey, surveyId, responses, surveyBaseUrl }: SummaryPageProps) => {
   const { selectedFilter, dateRange, resetState } = useResponseFilter();
-
   const searchParams = useSearchParams();
 
   useEffect(() => {
@@ -29,11 +30,11 @@ const ResponsePage = ({ environmentId, survey, surveyId, responses, surveyBaseUr
       resetState();
     }
   }, [searchParams]);
-
   // get the filtered array when the selected filter value changes
   const filterResponses: TResponse[] = useMemo(() => {
     return getFilterResponses(responses, selectedFilter, survey, dateRange);
   }, [selectedFilter, responses, survey, dateRange]);
+
   return (
     <ContentWrapper>
       <SummaryHeader
@@ -48,15 +49,11 @@ const ResponsePage = ({ environmentId, survey, surveyId, responses, surveyBaseUr
         survey={survey}
         totalResponses={responses}
       />
-      <SurveyResultsTabs activeId="responses" environmentId={environmentId} surveyId={surveyId} />
-      <ResponseTimeline
-        environmentId={environmentId}
-        surveyId={surveyId}
-        responses={filterResponses}
-        survey={survey}
-      />
+      <SurveyResultsTabs activeId="summary" environmentId={environmentId} surveyId={surveyId} />
+      <SummaryMetadata responses={filterResponses} survey={survey} />
+      <SummaryList responses={filterResponses} survey={survey} environmentId={environmentId} />
     </ContentWrapper>
   );
 };
 
-export default ResponsePage;
+export default SummaryPage;
